+++ src/Rarus/BonusServer/Cards/DTO/Fabric.php
@@ -9,6 +9,7 @@ use Money\Parser\DecimalMoneyParser;
 use Money\Currencies\ISOCurrencies;
 use Rarus\BonusServer\Cards\DTO\Level\LevelId;
 use Rarus\BonusServer\Users\DTO\UserId;
+use Rarus\BonusServer\Util\DateTimeParser;
 
 /**
  * Class Fabric
@@ -22,6 +23,7 @@ class Fabric
      * @param Currency $currency
      *
      * @return Card
+     * @throws \Rarus\BonusServer\Exceptions\ApiClientException
      */
     public static function initCardFromServerResponse(array $arCard, \Money\Currency $currency): Card
     {
@@ -40,7 +42,7 @@ class Fabric
             ->setCardStatus(Status\Fabric::initFromServerResponse($arCard));
 
         if ($arCard['date_last_transaction'] !== 0) {
-            $card->setDateLastTransaction(\DateTime::createFromFormat('U', (string)$arCard['date_last_transaction']));
+            $card->setDateLastTransaction(DateTimeParser::parseTimestampFromServerResponse((string)$arCard['date_last_transaction']));
         }
         if ($arCard['last_transaction'] !== '') {
             $card->setLastTransaction((string)$arCard['last_transaction']);